// src/lib/types/playerTypes.ts
/**
 * Типы данных, связанные с игроками
 */

/**
 * Интерфейс, описывающий игрока на сервере
 */
export interface Player {
  /** Уникальный идентификатор игрока */
  id: number;
  /** Имя игрока в формате Name_Surname */
  name: string;
  /** Уровень игрока в игре */
  level: number;
  /** Текущее здоровье игрока (0-100) */
  health: number;
  /** Текущая броня игрока (0-100) */
  armor: number;
  /** Количество денег у игрока */
  money: number;
  /** Игровые очки игрока */
  score: number;
  /** Пинг подключения (в миллисекундах) */
  ping: number;
  /** Время проведенное в игре в формате "XXч XXм" */
  playtime: string;
  /** IP-адрес игрока */
  ip: string;
  /** Имеет ли игрок админские права */
  isAdmin: boolean;
  /** Заглушен ли игрок (не может писать в чат) */
  isMuted: boolean;
  /** Находится ли игрок в тюрьме */
  isJailed: boolean;
}